requirements.txt
fastmcp>=2.0
psycopg2-binary>=2.9
uvicorn>=0.30
starlette>=0.40
